fixed the start of status effect systems.
mostly due to the nature of this unstructured test,  I didn't see that I started assigning the ids with 1 instead of 0


the update trick was done with a pre filtered signal in an internal event. however, the big problem stays that too many signals were fired and received by the agents.